create CRUD city

+++ app/Services/Country/DeleteService.php
@@ -1,9 +1,8 @@
 <?php
 namespace App\Services\Country;
 
-use App\Models\Country;
-use App\Validators\CountryValidator;
 use App\Repositories\CountryRepository;
+use App\Repositories\StateRepository;
 use App\Services\ServiceInterface;
 use Exception;
 use Illuminate\Support\Facades\DB;
@@ -24,6 +23,12 @@ final class DeleteService implements ServiceInterface
         $country = $countryRepository->find($data['id']);
         if($country==null){
             throw new Exception("Country not found.");
+        }
+
+        $stateRepository = new StateRepository();
+        $state = $stateRepository->findBy('country_id', $data['id']);
+        if($state!=null){
+            throw new Exception("Country in use.");
         }
         $country = $countryRepository->delete($data['id']);
         DB::commit();
