+++ app/Services/State/DeleteService.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Services\State;
 
+use App\Repositories\CityRepository;
 use App\Repositories\StateRepository;
 use App\Services\ServiceInterface;
 use Exception;
@@ -22,6 +23,12 @@ final class DeleteService implements ServiceInterface
         $state = $stateRepository->find($data['id']);
         if($state==null){
             throw new Exception("State not found.");
+        }
+
+        $cityRepository = new CityRepository();
+        $city = $cityRepository->findBy('state_id', $data['id']);
+        if($city!=null){
+            throw new Exception("State in use.");
         }
         $state = $stateRepository->delete($data['id']);
         DB::commit();
